requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
drf-spectacular==0.20.2
idna==3.3
inflection==0.5.1
jsonschema==4.2.1
pyrsistent==0.18.0
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
